Copy getImageStyles into iOS and Android forks
